file_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
Pillow>=10.3.0
python-magic==0.4.27
django-cors-headers==4.3.1
